Further refactored PlotEffect class and initialised test implementation
If you're failing to understand PlotEffect: it exists as a stack of
modifier arrays where the array on the top is ALWAYS the set of
modifiers imposed by the effect itself. All other arrays below it in the
stack contain the original production modifiers held by affected tiles,
allowing for effects to be reverted later on by calling the revert()
function. Effects can be imposed either by calling
[PlotEffect.impose(LandPlot, mode)] directly, or by setting up its
internal Runnable object to call that after executing other processes
and running the Runnable instead.

Note that the productionModifiers array in the LandPlot class, which was
previously of type int[], was changed to be of type float[] so that it
could be better integrated into the new LandPlot class.

+++ core/src/io/github/teamfractal/entity/LandPlot.java
@@ -71,7 +71,7 @@ public class LandPlot {
 	 * Saved modifiers for LandPlot.
 	 * [ Ore, Energy, Food ]
 	 */
-	int[] productionModifiers = {0, 0, 0};
+	float[] productionModifiers = {0, 0, 0};
 
 	/**
 	 * The base production amounts.
@@ -152,8 +152,8 @@ public class LandPlot {
 	 *
 	 * @return The amount of resources to be produced in an 2D array.
 	 */
-	public int[] produceResources() {
-		int[] produced = new int[3];
+	public float[] produceResources() {
+		float[] produced = new float[3];
 		for (int i = 0; i < 2; i++) {
 			produced[i] = productionAmounts[i] * productionModifiers[i];
 		}
@@ -165,7 +165,7 @@ public class LandPlot {
 	 * @param resource  The resource type to be calculated.
 	 * @return          Calculated amount of resource to be generated.
 	 */
-	public int produceResource(ResourceType resource) {
+	public float produceResource(ResourceType resource) {
 		if (this.hasRoboticon){
 			int resIndex = resourceTypeToIndex(resource);
 			return productionAmounts[resIndex] * productionModifiers[resIndex];
